Tried to fix bugs
Changed some methods to make the code work

This fixes #16
Tried to fix bugs

Changed some methods to make the code work

This fixes #16

+++ movieRating/core/src/main/java/core/UserHandling.java
@@ -49,7 +49,7 @@ public class UserHandling {
                     String movieTitle = movieAndRatingParts[0];
                     String genre = movieAndRatingParts[1];
                     String rating = movieAndRatingParts[2];
-                    user.rateMovie(new Movie(movieTitle, genre), Double.parseDouble(rating));
+                    user.rateMovie(new Movie(movieTitle, genre), Integer.parseInt(rating));
                 }
                 users.registerNewUser(user);
